CIRCLES-258: done doing that, tested with frontend to  check i didnt break anything
