+++ backend/server/server.py
@@ -1,6 +1,5 @@
 from fastapi import FastAPI
-from server.routers import api
-from server.routers import planner
+from server.routers import planner, courses, programs
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
@@ -32,7 +31,8 @@ app.add_middleware(
 )
 
 app.include_router(planner.router)
-app.include_router(api.router)
+app.include_router(courses.router)
+app.include_router(programs.router)
 
 @app.get('/')
 async def index():
