+++ frontend/src/pages/CourseSelector/SearchCourse.jsx
@@ -71,7 +71,7 @@ export default function SearchCourse() {
     console.log("new request");
     try {
       const res = await axios.post(
-        `http://localhost:8000/api/getAllUnlocked/`,
+        `http://localhost:8000/courses/getAllUnlocked/`,
         JSON.stringify(payload),
         {
           headers: {
