+++ frontend/src/pages/CourseSelector/courseMenu/CourseMenu.jsx
@@ -34,7 +34,7 @@ export default function CourseMenu() {
     // Local Development Testing
     const res = await axios.get("http://localhost:3000/structure.json");
     // Uncomment when DB is working
-    // const coreData = await axios.get(`http://localhost:8000/api/getCoreCourses/${programCode}/${specialisation}/${minor}`);
+    // const coreData = await axios.get(`http://localhost:8000/programs/getStructure/${programCode}/${specialisation}/${minor}`);
     dispatch(setStructure(res.data));
   };
   if (!structure) fetchProgression();
